rename clusters folder to cluster
